Moved cognito_util functions to backend/utils/cognito, removed cognito_util.py, and updated imports accordingly.
