+++ backend/__init__.py
@@ -2,7 +2,7 @@ from backend.models import Annotation, Project, Video, create_db_models, initial
 from backend.utils import get_environment_variable, parse_bool
 
 from .application import create_app, run_application
-from .cognito_util import sign_up, verify_sign_up
+from backend.utils.cognito import sign_up, verify_sign_up
 from .routes import (
     register_cognito_auth_endpoints,
     register_projects_endpoint,
